SCI: Use generated IDs for sound nodePtr
A hypothetical problem becomes a real problem.

After selecting a child at the start of MUMG CD, the game calls to
fade out the sound, zeroes out the longSong object's nodePtr, then
calls to kDoSoundInit again using the same longSong object while
the original sound is still fading out. If we use soundObj as the
nodePtr, this results in a situation where two sounds with the same
nodePtr are in the _sounds list at the same time, which is illegal
since then it is no longer possible to find the correct sound by
key. Instead, we now generate unique nodePtr values ourselves for
each new sound (just like how ScreenItem and ScrollWindow work).