SCI: Handle sample looping more accurately
In SCI0 SNDBLAST.DRV the number of loops would only be allowed to
go as low as 1, so this is not fully accurate in that regard, and
sounds would only loop on tick boundaries, so this may not have
ever actually been used since it wouldn't have sounded very good.
It may end up being the case that this functionality should just be
removed entirely because it was not ever used.